Moved read I/O classes to new ReadIO.hpp
